+++ src/main/java/org/total/spring/root/service/interfaces/StandingService.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface StandingService {
     public List<List<String>> getStandings(String seasonCode, String tournamentCode);
 
-    public List<Standing> getMatchDayStandings(String seasonCode, String tournamentCode, Byte matchDay);
+    public List<Standing> getMatchDayStandings(String seasonCode, String tournamentCode, Integer matchDay);
 
     public String getCachedStandings(String seasonCode, String tournamentCode);
 }
